requirements.txt
aioredis==1.3.1
alembic==1.6.5
asgiref==3.4.1
asyncio-redis==0.16.0
broadcaster==0.2.0
celery==5.1.2
factory-boy==3.2.0
fastapi==0.66.0
flower==1.0.0
gunicorn==20.1.0
Jinja2==3.0.1
Pillow==8.3.1
psycopg2-binary==2.9.1
pytest==6.2.4
pytest-cov==2.12.1
pytest-factoryboy==2.1.0
python-multipart==0.0.5
python-socketio==5.3.0
redis==3.5.3
requests==2.25.1
SQLAlchemy==1.4.20
uvicorn[standard]==0.14.0
watchgod==0.7
